fix ln error

+++ script/build-mingw-toolchain.sh
@@ -34,7 +34,7 @@ if [[ "$_THREAD" == "mcf" ]]; then
   "$_SCRIPT_DIR/mingw/mcf.sh"
 fi
 
-ln -s "$_PKG_DIR$_PREFIX" "$_PREFIX"
+[[ -L "$_PREFIX" ]] || ln -s "$_PKG_DIR$_PREFIX" "$_PREFIX"
 
 title "Building MinGW libiconv"
 "$_SCRIPT_DIR/mingw/iconv.sh"
